AccountSettingsPageのコンテンツを遅延読み込みするようにした
