+++ src/Beutl/Pages/SettingsPages/SignInScreen.axaml
@@ -1,0 +1,89 @@
+<UserControl x:Class="Beutl.Pages.SettingsPages.SignInScreen"
+             xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:lang="using:Beutl.Language"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:ui="using:FluentAvalonia.UI.Controls"
+             xmlns:vm="using:Beutl.ViewModels.SettingsPages"
+             d:DesignHeight="400"
+             d:DesignWidth="300"
+             x:DataType="vm:AccountSettingsPageViewModel"
+             RenderOptions.BitmapInterpolationMode="HighQuality"
+             mc:Ignorable="d">
+    <StackPanel>
+        <TextBlock Margin="0,0,0,24"
+                   Text="{x:Static lang:SettingsPage.SignIn}"
+                   Theme="{StaticResource SubtitleTextBlockStyle}" />
+        <TextBlock Margin="0,0,0,8"
+                   Text="{x:Static lang:SettingsPage.SignInWith}"
+                   Theme="{StaticResource BodyStrongTextBlockStyle}" />
+        <Grid ColumnDefinitions="*,16,*">
+
+            <Button Height="48"
+                    HorizontalAlignment="Stretch"
+                    HorizontalContentAlignment="Center"
+                    Command="{CompiledBinding SignInWithGoogle}">
+                <Grid ColumnDefinitions="48,*">
+                    <Svg Height="48" Path="/Assets/social/btn_google_light_normal_ios.svg" />
+
+                    <TextBlock Grid.Column="1"
+                               VerticalAlignment="Center"
+                               FontSize="16"
+                               Text="Google" />
+                </Grid>
+            </Button>
+            <Button Grid.Column="2"
+                    Height="48"
+                    HorizontalAlignment="Stretch"
+                    HorizontalContentAlignment="Center"
+                    Command="{CompiledBinding SignInWithGitHub}">
+                <Grid ColumnDefinitions="48,*">
+                    <Image x:Name="githubLightLogo"
+                           Height="24"
+                           IsVisible="False"
+                           Source="/Assets/social/GitHub-Mark-120px-plus.png" />
+                    <Image x:Name="githubDarkLogo"
+                           Height="24"
+                           IsVisible="False"
+                           Source="/Assets/social/GitHub-Mark-Light-120px-plus.png" />
+
+                    <TextBlock Grid.Column="1"
+                               VerticalAlignment="Center"
+                               FontSize="16"
+                               Text="GitHub" />
+                </Grid>
+            </Button>
+        </Grid>
+
+        <TextBlock MinHeight="16"
+                   Margin="0,16,0,0"
+                   Classes="error"
+                   Text="{CompiledBinding Error.Value}" />
+
+        <Button Height="40"
+                Margin="0,8,0,0"
+                HorizontalAlignment="Stretch"
+                Classes="accent"
+                Command="{CompiledBinding SignIn}"
+                Content="{x:Static lang:SettingsPage.SignInViaBrowser}" />
+
+        <Grid Margin="0,16,0,0" ColumnDefinitions="*,*">
+            <ui:HyperlinkButton Grid.Column="0"
+                                HorizontalAlignment="Center"
+                                Content="{x:Static lang:Strings.PrivacyPolicy}"
+                                NavigateUri="https://beutl.beditor.net/Privacy" />
+            <ui:HyperlinkButton Grid.Column="1"
+                                HorizontalAlignment="Center"
+                                Content="{x:Static lang:Strings.TermsOfService}"
+                                NavigateUri="https://beutl.beditor.net/TermsOfService" />
+        </Grid>
+
+        <Button Height="40"
+                Margin="0,16,0,0"
+                HorizontalAlignment="Stretch"
+                Classes.transparent="{CompiledBinding !SigningIn.Value}"
+                Command="{CompiledBinding Cancel}"
+                Content="{x:Static lang:Strings.Cancel}" />
+    </StackPanel>
+</UserControl>
